requirements.txt
flask-restx==0.1.1
Flask==1.1.1
requests==2.22.0
python-dotenv==0.10.3
flask-cors==3.0.8